3rd and last step to implements hibernate

+++ server/src/main/java/com/aion/server/service/infra/exception/ShopInsertionException.java
@@ -3,7 +3,7 @@ package com.aion.server.service.infra.exception;
 public class ShopInsertionException extends ShopException {
 
     public ShopInsertionException(final long itemId,
-                                  final int userId) {
+                                  final long userId) {
 
         super("Failed to insert item " + itemId + " for user " + userId);
     }
