+++ server/src/main/java/com/aion/server/service/infra/exception/ShopPerformPurchaseException.java
@@ -3,7 +3,7 @@ package com.aion.server.service.infra.exception;
 public class ShopPerformPurchaseException extends ShopException {
 
     public ShopPerformPurchaseException(final long itemId,
-                                        final int userId) {
+                                        final long userId) {
 
         super("Failed to perform purchase user : " + userId + " item" + itemId);
     }
